End of Restricted Pages

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -7,14 +9,10 @@ from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
-
-# rooms = [
-#     {'id':1, 'name':'Lets learn Python!'},
-#     {'id':2, 'name':'Design with me'},
-#     {'id':3, 'name':'Frontend developers'},
-# ]
-
 def loginPage(request):
+    
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -36,6 +34,7 @@ def loginPage(request):
     context = {}
     return render(request, 'base/login_register.html', context)
 
+
 def logoutUser(request):
     logout(request)
     return redirect('home')
@@ -53,11 +52,14 @@ def home(request):
     context = {'rooms':rooms, 'topics':topics, 'room_count':room_count}
     return render(request, 'base/home.html', context)
 
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room':room}        
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     
@@ -71,10 +73,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('Only the creator of the room can edit or delete the room.')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -86,8 +91,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('Only the creator of the room can edit or delete the room.')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
